CLR proxy handles are now long

+++ jni4net.j/src/main/java/java_/io/Closeable_.java
@@ -29,7 +29,7 @@ public final class Closeable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Closeable extends system.Object implements java.io.Closeable {
     
-    protected __Closeable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Closeable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
