+++ jni4net.j/src/main/java/java_/io/Flushable_.java
@@ -29,7 +29,7 @@ public final class Flushable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Flushable extends system.Object implements java.io.Flushable {
     
-    protected __Flushable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Flushable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
