+++ jni4net.j/src/main/java/java_/lang/reflect/AnnotatedElement_.java
@@ -29,7 +29,7 @@ public final class AnnotatedElement_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __AnnotatedElement extends system.Object implements java.lang.reflect.AnnotatedElement {
     
-    protected __AnnotatedElement(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __AnnotatedElement(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
