+++ jni4net.j/src/main/java/java_/lang/reflect/Member_.java
@@ -29,7 +29,7 @@ public final class Member_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Member extends system.Object implements java.lang.reflect.Member {
     
-    protected __Member(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Member(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
