+++ jni4net.j/src/main/java/java_/lang/reflect/ParameterizedType_.java
@@ -29,7 +29,7 @@ public final class ParameterizedType_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __ParameterizedType extends system.Object implements java.lang.reflect.ParameterizedType {
     
-    protected __ParameterizedType(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __ParameterizedType(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
