+++ jni4net.j/src/main/java/java_/util/Enumeration_.java
@@ -29,7 +29,7 @@ public final class Enumeration_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Enumeration extends system.Object implements java.util.Enumeration {
     
-    protected __Enumeration(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Enumeration(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
