+++ jni4net.j/src/main/java/java_/util/List_.java
@@ -29,7 +29,7 @@ public final class List_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __List extends system.Object implements java.util.List {
     
-    protected __List(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __List(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
