+++ jni4net.j/src/main/java/java_/util/Map_.java
@@ -29,7 +29,7 @@ public final class Map_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Map extends system.Object implements java.util.Map {
     
-    protected __Map(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Map(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
