+++ jni4net.j/src/main/java/system/Array.java
@@ -15,7 +15,7 @@ public class Array extends system.Object implements system.ICloneable, system.co
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected Array(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected Array(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
