+++ jni4net.j/src/main/java/system/Exception.java
@@ -23,22 +23,22 @@ package system;
 	@net.sf.jni4net.attributes.ClrType
 public class Exception extends java.lang.RuntimeException implements IClrProxy, system.IObject {
 
-	private int clrHandle;
+	private long clrHandle;
 
 	@Override
 	public java.lang.String toString() {
 		return ToString();
 	}
 
-	public int getClrHandle() {
+	public long getClrHandle() {
 		return clrHandle;
 	}
 
-	public void initProxy(int handle) {
+	public void initProxy(long handle) {
 		clrHandle = handle;
 	}
 
-	protected Exception(net.sf.jni4net.inj.INJEnv env, int handle) {
+	protected Exception(net.sf.jni4net.inj.INJEnv env, long handle) {
 		clrHandle = handle;
 	}
