+++ jni4net.j/src/main/java/system/Guid.java
@@ -15,7 +15,7 @@ public class Guid extends system.ValueType implements system.IFormattable, syste
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected Guid(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected Guid(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
