+++ jni4net.j/src/main/java/system/ICloneable_.java
@@ -29,7 +29,7 @@ public final class ICloneable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __ICloneable extends system.Object implements system.ICloneable {
     
-    protected __ICloneable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __ICloneable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
