+++ jni4net.j/src/main/java/system/IObject_.java
@@ -40,7 +40,7 @@ public final class IObject_ {
  */
 @net.sf.jni4net.attributes.ClrProxy
 class __IObject extends system.Object implements system.IObject {
-	protected __IObject(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+	protected __IObject(net.sf.jni4net.inj.INJEnv __env, long __handle) {
 			super(__env, __handle);
 	}
 }
