+++ jni4net.j/src/main/java/system/MarshalByRefObject.java
@@ -15,7 +15,7 @@ public class MarshalByRefObject extends system.Object {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected MarshalByRefObject(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected MarshalByRefObject(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
