+++ jni4net.j/src/main/java/system/NotSupportedException.java
@@ -15,7 +15,7 @@ public class NotSupportedException extends system.SystemException {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected NotSupportedException(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected NotSupportedException(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
