+++ jni4net.j/src/main/java/system/collections/IDictionary_.java
@@ -29,7 +29,7 @@ public final class IDictionary_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __IDictionary extends system.Object implements system.collections.IDictionary {
     
-    protected __IDictionary(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __IDictionary(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
