+++ jni4net.j/src/main/java/system/io/FileStream.java
@@ -15,7 +15,7 @@ public class FileStream extends system.io.Stream {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected FileStream(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected FileStream(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
