+++ jni4net.j/src/main/java/system/io/IOException.java
@@ -15,7 +15,7 @@ public class IOException extends system.SystemException {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected IOException(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected IOException(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
