+++ jni4net.j/src/main/java/system/reflection/ConstructorInfo.java
@@ -15,7 +15,7 @@ public class ConstructorInfo extends system.reflection.MethodBase {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected ConstructorInfo(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected ConstructorInfo(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
