+++ jni4net.j/src/main/java/system/reflection/FieldInfo.java
@@ -15,7 +15,7 @@ public class FieldInfo extends system.reflection.MemberInfo {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected FieldInfo(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected FieldInfo(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
