+++ jni4net.j/src/main/java/system/reflection/MethodBase.java
@@ -15,7 +15,7 @@ public class MethodBase extends system.reflection.MemberInfo {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected MethodBase(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected MethodBase(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
