+++ jni4net.j/src/main/java/system/runtime/serialization/ISerializable_.java
@@ -29,7 +29,7 @@ public final class ISerializable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __ISerializable extends system.Object implements system.runtime.serialization.ISerializable {
     
-    protected __ISerializable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __ISerializable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
