+++ jni4net.tested.j/src/main/java/net/sf/jni4net/tested/CInterfacesHelper.java
@@ -15,7 +15,7 @@ public class CInterfacesHelper extends system.Object {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected CInterfacesHelper(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected CInterfacesHelper(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
