+++ jni4net.tested.j/src/main/java/net/sf/jni4net/tested/StaticMethods.java
@@ -15,7 +15,7 @@ public class StaticMethods extends system.Object {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected StaticMethods(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected StaticMethods(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
